app/requirements.txt
cmsplugin-filer>=0.10.2
cmsplugin-sections
dj-database-url==0.4.1
Django==1.8.13
django-classy-tags==0.7.2
django-cms==3.2.5
django-extensions==1.6.1
django-filer>=0.9.12
django-formtools==1.0
django-mptt==0.8.4
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.1.1
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.8
djangocms-link==1.7.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.3
djangocms-video==1.0.0
easy_thumbnails
html5lib==0.9999999
Pillow==3.2.0
pytz==2016.4
six==1.10.0
tzlocal==1.2.2
mysqlclient==1.3.7
Werkzeug==0.11.7
